refactor: tenderdash smoke tests
