+++ test/smoke/index.spec.js
@@ -2,7 +2,7 @@
 describe('Smoke', () => {
   require('./core');
   require('./drive');
-  require('./tendermint');
+  require('./tenderdash');
   require('./dapi');
   require('./sentinel');
   require('./faucet');
